+++ demisto_sdk/commands/common/errors.py
@@ -1789,7 +1789,7 @@ class Errors:
     @staticmethod
     @error_code_decorator
     def invalid_readme_relative_url_error(path):
-        return f'Relative url is not supported within README, If this is not a relative url make sure you added ' \
+        return f'Relative urls are not supported within README, If this is not a relative url make sure you add ' \
                f'https:// at the start of it:\n{path}. '
 
     @staticmethod