+++ demisto_sdk/commands/common/tests/pack_unique_files_test.py
@@ -655,18 +655,23 @@ class TestPackUniqueFilesValidator:
                     - Validation fails
                     - Invalid relative paths were caught correctly
                     - Invalid absolute paths were not caught.
+                    - Image paths were not caught
         """
         self.validator = PackUniqueFilesValidator(os.path.join(self.FILES_PATH, 'DummyPack2'))
 
         result = self.validator.validate_pack_readme_relative_urls()
         errors = self.validator.get_errors()
         assert not result
-        assert 'Relative url is not supported within README, If this is not a relative url make sure you added ' \
+        assert 'Relative urls are not supported within README, If this is not a relative url make sure you add ' \
                'https:// at the start of it:\nrelative1.com. ' in errors
-        assert 'Relative url is not supported within README, If this is not a relative url make sure you added ' \
+        assert 'Relative urls are not supported within README, If this is not a relative url make sure you add ' \
                'https:// at the start of it:\nwww.relative2.com. ' in errors
+        assert 'Relative urls are not supported within README, If this is not a relative url make sure you add ' \
+               'https:// at the start of it:\nhreftesting.com' in errors
         # this path is not an image path and should not be shown.
         assert 'htttps://www.good.co.il' not in errors
+        assert 'https://example.com' not in errors
+        assert 'doc_files/High_Risk_User.png' not in errors
 
     @pytest.mark.parametrize('readme_content, is_valid', [
         ('Hey there, just testing', True),